# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Smart Goal Breaker API","description":"AI-powered goal breakdown service that transforms vague goals into actionable steps","version":"1.0.0"},"servers":[{"url":"http://localhost:8001","description":"Development server"},{"url":"http://localhost:8000","description":"Production server"}],"paths":{"/api/goals":{"post":{"tags":["goals"],"summary":"Create a new goal","description":"Break down a goal into 5 actionable subtasks with complexity scores using AI","operationId":"create_goal_api_goals_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoalCreate"}}}},"responses":{"200":{"description":"Goal successfully created with AI-generated subtasks","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Goal"},"example":{"id":1,"description":"Launch a startup","created_at":"2025-11-26T00:00:00Z","subtasks":[{"description":"Validate business idea","complexity_score":6},{"description":"Create business plan","complexity_score":7}]}}}},"500":{"description":"AI generation failed or database error"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["goals"],"summary":"Get all goals","description":"Retrieve a list of all goals with their subtasks","operationId":"read_goals_api_goals_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"List of goals retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Goal"},"title":"Response Read Goals Api Goals Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Goal":{"properties":{"description":{"type":"string","title":"Description"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"subtasks":{"items":{"$ref":"#/components/schemas/SubTask"},"type":"array","title":"Subtasks","default":[]}},"type":"object","required":["description","id","created_at"],"title":"Goal"},"GoalCreate":{"properties":{"description":{"type":"string","title":"Description"}},"type":"object","required":["description"],"title":"GoalCreate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SubTask":{"properties":{"description":{"type":"string","title":"Description"},"complexity_score":{"type":"integer","title":"Complexity Score"},"id":{"type":"integer","title":"Id"},"goal_id":{"type":"integer","title":"Goal Id"}},"type":"object","required":["description","complexity_score","id","goal_id"],"title":"SubTask"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"goals","description":"Operations for managing and breaking down goals"}]}